+++ javax/src/main/java/me/panpf/javax/util/IndexedAction2.java
@@ -18,7 +18,6 @@ package me.panpf.javax.util;
 
 import org.jetbrains.annotations.NotNull;
 
-public interface Transformer2<T1, T2, V> {
-    @NotNull
-    V transform(@NotNull T1 t1, @NotNull T2 t2);
+public interface IndexedAction2<T1, T2> {
+    void action(int index, @NotNull T1 t1, @NotNull T2 t2);
 }